pyproject.toml
[project]
name = "test-mcp-task"
version = "0.1.0"
description = "Тестовое задание с MCP-сервером для получения информации о продуктах компании"
readme = "README.md"
requires-python = ">=3.11"
authors = ["Alex Karav <sashagameac@gmail.com>"]
dependencies = [
    "apscheduler>=3.11.0",
    "bs4>=0.0.2",
    "google-api-python-client>=2.183.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp>=1.15.0",
    "mcpo>=0.0.17",
    "openpyxl>=3.1.5",
    "pydantic>=2.11.9",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
]
